Fix type Float64

+++ src/greedy.jl
@@ -14,7 +14,7 @@
 Takes a set of candidates and removes all elements with bad criterion.
 """
 function filter_on_max_criterion(candidates::Array{Int,1},
-                                 criterion::Union{Array{Int,1},Array{Float,1}})
+                                 criterion::Union{Array{Int,1},Array{Float64,1}})
     tmp_candidates = [candidates[1]]
     max_criterion = criterion[1]
     for i in 2:length(criterion)
